# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MYE030 Data Integration & Visualization API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/authors": {
      "get": {
        "tags": [
          "authors"
        ],
        "summary": "Search Authors",
        "description": "Return a paginated author listing, optionally filtered by name substring.\n\nBy default, only authors with at least one indexed contribution are\nreturned. Pass ``min_articles=0`` to include silent authors (i.e.\nnames present in ``lookup_authors`` but never matched to any article).\n\nSortable columns: ``author_name``, ``author_id``, ``total_articles``,\n``earliest_year``, ``latest_year``.",
        "operationId": "search_authors_api_authors_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          },
          {
            "name": "name_query",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name Query"
            }
          },
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order By"
            }
          },
          {
            "name": "order_dir",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(asc|desc)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order Dir"
            }
          },
          {
            "name": "min_articles",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 1,
              "title": "Min Articles"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_AuthorSummary_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/authors/{author_id}": {
      "get": {
        "tags": [
          "authors"
        ],
        "summary": "Get Author Profile",
        "description": "Return the aggregated profile of one author.",
        "operationId": "get_author_profile_api_authors__author_id__get",
        "parameters": [
          {
            "name": "author_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Author Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorProfile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/authors/{author_id}/yearly-statistics": {
      "get": {
        "tags": [
          "authors"
        ],
        "summary": "Get Author Yearly Statistics",
        "description": "Return the per-year article counts for the author's line chart.\n\nTrims a trailing partial year by default.",
        "operationId": "get_author_yearly_statistics_api_authors__author_id__yearly_statistics_get",
        "parameters": [
          {
            "name": "author_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Author Id"
            }
          },
          {
            "name": "include_partial_last_year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Partial Last Year"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuthorYearlyStatistic"
                  },
                  "title": "Response Get Author Yearly Statistics Api Authors  Author Id  Yearly Statistics Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/authors/{author_id}/articles": {
      "get": {
        "tags": [
          "authors"
        ],
        "summary": "List Author Articles",
        "description": "Return a paginated list of every article the author has co-authored.\n\nSortable columns: ``title``, ``year``, ``venue_title``, ``venue_type``.",
        "operationId": "list_author_articles_api_authors__author_id__articles_get",
        "parameters": [
          {
            "name": "author_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Author Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          },
          {
            "name": "start_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 1900
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Year"
            }
          },
          {
            "name": "end_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 1900
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Year"
            }
          },
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order By"
            }
          },
          {
            "name": "order_dir",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(asc|desc)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order Dir"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_AuthorArticle_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuthorArticle": {
        "properties": {
          "article_id": {
            "type": "integer",
            "title": "Article Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "venue_type": {
            "type": "string",
            "title": "Venue Type"
          },
          "venue_id": {
            "type": "integer",
            "title": "Venue Id"
          },
          "venue_title": {
            "type": "string",
            "title": "Venue Title"
          }
        },
        "type": "object",
        "required": [
          "article_id",
          "title",
          "venue_type",
          "venue_id",
          "venue_title"
        ],
        "title": "AuthorArticle",
        "description": "Single article authored by the user, regardless of venue type."
      },
      "AuthorProfile": {
        "properties": {
          "author_id": {
            "type": "integer",
            "title": "Author Id"
          },
          "author_name": {
            "type": "string",
            "title": "Author Name"
          },
          "total_articles": {
            "type": "integer",
            "title": "Total Articles"
          },
          "earliest_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Earliest Year"
          },
          "latest_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Year"
          },
          "average_articles_per_year": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Articles Per Year"
          }
        },
        "type": "object",
        "required": [
          "author_id",
          "author_name",
          "total_articles"
        ],
        "title": "AuthorProfile",
        "description": "Full profile aggregating an author's publication record."
      },
      "AuthorSummary": {
        "properties": {
          "author_id": {
            "type": "integer",
            "title": "Author Id"
          },
          "author_name": {
            "type": "string",
            "title": "Author Name"
          },
          "total_articles": {
            "type": "integer",
            "title": "Total Articles",
            "default": 0
          },
          "earliest_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Earliest Year"
          },
          "latest_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Year"
          }
        },
        "type": "object",
        "required": [
          "author_id",
          "author_name"
        ],
        "title": "AuthorSummary",
        "description": "Lightweight projection used by search and list endpoints.\n\nIncludes aggregate counts so the browse table can rank authors by\noutput without the caller having to follow a profile link."
      },
      "AuthorYearlyStatistic": {
        "properties": {
          "year": {
            "type": "integer",
            "title": "Year"
          },
          "articles_count": {
            "type": "integer",
            "title": "Articles Count"
          }
        },
        "type": "object",
        "required": [
          "year",
          "articles_count"
        ],
        "title": "AuthorYearlyStatistic",
        "description": "One bucket of an author's per-year publication count."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_AuthorArticle_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuthorArticle"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Page Size"
          },
          "total_items": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Items"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total_items"
        ],
        "title": "PaginatedResponse[AuthorArticle]"
      },
      "PaginatedResponse_AuthorSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuthorSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Page Size"
          },
          "total_items": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Items"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total_items"
        ],
        "title": "PaginatedResponse[AuthorSummary]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}